+++ src/procheader.h
@@ -43,14 +43,14 @@ struct _Header
 };
 
 gint procheader_get_one_field		(gchar		*buf,
-					 gint		 len,
+					 size_t		 len,
 					 FILE		*fp,
 					 HeaderEntry	 hentry[]);
 gint procheader_get_one_field_asis	(gchar		*buf,
-					 gint		 len,
+					 size_t		 len,
 					 FILE		*fp);
 gchar *procheader_get_unfolded_line	(gchar		*buf,
-					 gint		 len,
+					 size_t		 len,
 					 FILE		*fp);
 
 #if 0
